Add fix to remove tmp file only if it exists

+++ faas_profiler_python/profiler.py
@@ -157,14 +157,14 @@ class Profiler(Loggable):
 
         self._start_capturing()
         self._start_default_measurements()
-        self._start_periodic_measurements()
+        # self._start_periodic_measurements()
 
     def stop(self) -> None:
         """
         Stops the profiling.
         """
         self.logger.info("Profile run stopped.")
-        self._stop_periodic_measurements()
+        # self._stop_periodic_measurements()
         self._stop_default_measurements()
         self._stop_capturing()
 
@@ -206,9 +206,10 @@ class Profiler(Loggable):
                 self.logger.error(
                     f"Exporting with {exporter_plugin.cls} failed: {err}")
 
-        self.logger.info(
-            f"[EXPORT]: Delete tmp storage file: {self.periodic_results_path}")
-        os.remove(self.periodic_results_path)
+        if os.path.exists(self.periodic_results_path):
+            self.logger.info(
+                f"[EXPORT]: Delete tmp storage file: {self.periodic_results_path}")
+            os.remove(self.periodic_results_path)
 
     def _start_default_measurements(self):
         """
